Fake Physics routine is added but not worked yet due to some bugs.
git-svn-id: http://svn.sourceforge.jp/svnroot/mikumikudroid/trunk@21 fc36963f-cfbd-4723-99f8-9c1cee413372

+++ MikuMikuDroid/src/jp/gauzau/MikuMikuDroid/RigidBody.java
@@ -17,7 +17,7 @@ public class RigidBody {
 	public float friction;
 	public byte type;
 	public float[] cur_location;
-	public float[] cur_rotation;
-	public float[] cur_v;
+	public double[] cur_rotation;
+	public double[] cur_v;
 
 }
